pyproject.toml
[project]
name = "psx-scraper"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiohttp>=3.11.10",
    "asyncio>=3.4.3",
    "beautifulsoup4>=4.12.3",
    "lxml>=5.3.0",
    "python-dotenv>=1.0.1",
    "requests>=2.32.3",
]

[dependency-groups]
dev = [
    "ipykernel>=6.29.5",
]
